service/requirements.txt
aniso8601==6.0.0
attrs==19.1.0
Click==7.0
Flask==1.0.2
Flask-Cors==3.0.7
flask-restplus==0.12.1
itsdangerous==1.1.0
Jinja2==2.10.1
jsonschema==3.0.1
MarkupSafe==1.1.1
numpy==1.16.2
pyrsistent==0.14.11
pytz==2019.1
scikit-learn==0.20.3
scipy==1.2.1
six==1.12.0
sklearn==0.0
Werkzeug==0.15.2
seaborn==0.10.0
pandas==1.0.1
matplotlib==3.1.3
tensorflow-hub==0.7.0
tensorflow==2.1.0
absl-py==0.9.0

